Job_Finder_using_strands/requirements.txt
strands-agents[gemini]>=1.30.0
strands-agents-tools>=0.1.0

# Resume parsing
pypdf>=4.0.0
python-docx>=1.1.0

# Scraping
playwright>=1.46.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# UI
streamlit>=1.37.0

# Data + config
pydantic>=2.8.0
pyyaml>=6.0.1
python-dotenv>=1.0.1

# Utilities
rich>=13.7.0
tenacity>=9.0.0
